refactor: update import paths for User, Company, and JobListing types to use models directory
