+++ tests/routes/jobListings.test.ts
@@ -3,8 +3,8 @@ import { CompanyModel } from "@/models/Company";
 import { InterviewSessionModel } from "@/models/InterviewSession";
 import { JobListingModel } from "@/models/JobListing";
 import { UserModel } from "@/models/User";
-import type { Company } from "@/types/Company";
-import type { JobListing } from "@/types/JobListing";
+import type { Company } from "@/types/models/Company";
+import type { JobListing } from "@/types/models/JobListing";
 import type { Express } from "express";
 import jwt from "jsonwebtoken";
 import mongoose from "mongoose";
